# Fixed the pivot selection in the group quick-sort