--- a/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/storage/ldb/ArrayGroupSort.java
+++ b/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/storage/ldb/ArrayGroupSort.java
@@ -61,11 +61,13 @@ public class ArrayGroupSort {
     }
 
     private int partition(long[] array, int low, int high) {
-        int pivotIdx = high;
+        int mid = low + (high-low) / 2;
+        swap(array, mid, high);
+
         int i = low;
 
         for (int j = low; j < high; j += groupSize) {
-            if (isLess(array, j, pivotIdx)) {
+            if (isLess(array, j, high)) {
                 swap(array, j, i);
                 i += groupSize;
             }