--- a/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/storage/ldb/ArrayGroupSort.java
+++ b/bookkeeper-server/src/main/java/org/apache/bookkeeper/bookie/storage/ldb/ArrayGroupSort.java
@@ -61,7 +61,7 @@ public class ArrayGroupSort {
     }
 
     private int partition(long[] array, int low, int high) {
-        int mid = low + (high-low) / 2;
+        int mid = low + (high - low) / 2;
         swap(array, mid, high);
 
         int i = low;