### Motivation

We have a custom implementation of quick-sort in the `WriteCache` that is used to sort an array of long that are grouped in 4: (ledgerId, entryId, offset, length). 

Because of the choice of the partition pivot, the sort results in very poor performance where the items are already pre-sorted, or mostly-sorted. Similarly, it happens when they are reverse-sorted. One of the side effects are also stack-overflow when the array is big and pre-sorted. 

The problem is that we're always picking the last item in the segment as the pivot, leading to always the worst case partitioning (1, N-1) items. 

The simplest solution is to pick the mid item as pivot, which works well for sorted, reverse and random scenarios. Later we can try more sophisticated approaches, such as the ones from `Arrays.sort()`.

### Benchmarks

I've compared the different cases with the baseline of `Arrays.sort()`:

Before:

```
Benchmark                                    Mode  Cnt       Score       Error  Units
GroupSortBenchmark.preSortedArraySort       thrpt    3  140157.757 ± 16108.997  ops/s
GroupSortBenchmark.preSortedGroupSort       thrpt    3       6.278 ±     0.149  ops/s
GroupSortBenchmark.randomArraySort          thrpt    3     647.256 ±    75.014  ops/s
GroupSortBenchmark.randomGroupSort          thrpt    3    1465.937 ±    80.175  ops/s
GroupSortBenchmark.reverseSortedArraySort   thrpt    3  138949.197 ± 10135.254  ops/s
GroupSortBenchmark.reverseSortedGroupSort   thrpt    3       6.301 ±     1.130  ops/s
```

After:

```
Benchmark                                   Mode  Cnt       Score      Error  Units
GroupSortBenchmark.preSortedArraySort      thrpt    3  138807.541 ± 3275.438  ops/s
GroupSortBenchmark.preSortedGroupSort      thrpt    3    1978.380 ±  106.142  ops/s
GroupSortBenchmark.randomArraySort         thrpt    3     673.768 ±   79.817  ops/s
GroupSortBenchmark.randomGroupSort         thrpt    3    1172.314 ±   18.191  ops/s
GroupSortBenchmark.reverseSortedArraySort  thrpt    3  141224.984 ± 2783.076  ops/s
GroupSortBenchmark.reverseSortedGroupSort  thrpt    3    1958.383 ±   73.815  ops/s
```

The `preSortedGroupSort` goes from 6.2 to 1978.3 ops/s.